+++ test/integration/gen_class/integration_test_run3.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-# Integration testing program for the arg_parser.py module.
 # This will run all the units tests for this program.
 # Will need to run this from the base directory where the module file 
 #   is located at.